Fixed offset issue
