+++ src/index.css
@@ -31,3 +31,6 @@ code {
   width: 550px;
   word-break: break-all;
 }
+.header {
+  height: 435px;
+}
